Adding and updating the input argments in 'start_execution.py'
Adding more flexibility. Now the paths to the inputs, outputs, and python libraries
can all be specified, along with timestep, and start and end times.  It also
doesn't matter if you call from the same directory that it is located in or not.

Also, adding option to read in a JSON file with all the arguments.

Making sure we use os.path.join in places that use io_dir

v.5.1